requirements.txt
aiohttp==3.7.4.post0
aniso8601==7.0.0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2
django-channels==0.7.0
django-channels-graphql-ws==0.8.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
msgpack==1.0.2
multidict==5.1.0
oauthlib==3.1.0
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
Twisted
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
zope.interface
